sync with spring source

+++ spring-test/src/main/java/org/springframework/test/web/reactive/server/HeaderAssertions.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -181,14 +181,14 @@ public class HeaderAssertions {
 	/**
 	 * Expect an "Expires" header with the given value.
 	 */
-	public WebTestClient.ResponseSpec expires(int expires) {
+	public WebTestClient.ResponseSpec expires(long expires) {
 		return assertHeader("Expires", expires, getHeaders().getExpires());
 	}
 
 	/**
 	 * Expect a "Last-Modified" header with the given value.
 	 */
-	public WebTestClient.ResponseSpec lastModified(int lastModified) {
+	public WebTestClient.ResponseSpec lastModified(long lastModified) {
 		return assertHeader("Last-Modified", lastModified, getHeaders().getLastModified());
 	}
